refactor: restructure retirement tax calculations and address review feedback
