+++ config/tax.json
@@ -43,7 +43,17 @@
                 ]
             },
             "StandardTaxDeduction": 14600,
-            "JointTaxDeduction": 27700
+            "JointTaxDeduction": 27700,
+            "CapitalGainsTax": {
+                "Individual": {
+                    "LowerBounds": [0, 47026, 518901],
+                    "Percents": [0.0, 0.15, 0.2]
+                },
+                "Joint": {
+                    "LowerBounds": [0, 94051, 583751],
+                    "Percents": [0.0, 0.15, 0.2]
+                }
+            }
         },
         "StateTax": {
             "California": {
